Move code into src, add setuptools config for pip install
