+++ Pipfile
@@ -8,6 +8,7 @@ pyrtlsdr = {extras = ["lib"], version = "*"}
 numpy = "*"
 matplotlib = "*"
 scipy = "*"
+kiwitracker = {editable = true, path = "."}
 
 [dev-packages]
