+++ setup.cfg
@@ -1,0 +1,16 @@
+[metadata]
+name = KiwiTracker
+version = 0.0.1
+
+[options]
+package_dir=
+    =src
+packages = find:
+install_requires =
+    numpy
+    scipy
+    pyrtlsdr
+
+[options.packages.find]
+where = src
+exclude = tests
